requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
opencv-python>=4.5.0
timm>=0.5.0
tensorboard>=2.7.0